// src/pages/package-A/customer/user-enterprise/types.ts
import Actions from './actions';
import {EnterpriseInfoVO} from 'api/enterpriseInfoBaseController';

export interface IMainReducer {
  isReady: boolean;
  isLoading?: boolean;
  enterpriseNum: '';
  industry: '';
  enterpriseNumValue: number;
  industryValue: number;
  enterpriseInfoVO: EnterpriseInfoVO;
}

export type ActionType = ReturnType<typeof Actions>;
export type IAllReducerProps = {
  main: IMainReducer;

  [name: string]: any;
};

//默认是全部的属性,可以自定义
export type IProps = IAllReducerProps & ActionType;

export type IInfoProps = {};
export type IInfoState = {};

//create by moon https://github.com/creasy2010/moon